Enemy AI and animations
put AI onto the turret enemy so that it tracks the player when you get within range of it.

Added in animations to show each state of both the walker and turret. Each state looks different so you can tell what state it is in.

Issues:
the player weapon wont shoot anymore even though code is correct.

Similar code issues with getting the walker enemy to change state when you get with in range of it